pyproject.toml
[project]
name = "django-esi-auth"
version = "0.1.0"
description = "Eve Online login using django-esi"
authors = [
    {name = "Tim Davis",email = "binary.god@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5,<6)",
    "django-esi (>=6.0.0,<7.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [
    {include = "django_esi_auth"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
